- Update in-memory plugin
