+++ plugins/memory/config.go
